% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptloc.R
\name{script_file_get}
\alias{script_file_get}
\title{Get location of script that was executed through Rscript}
\usage{
script_file_get(cargs)
}
\arguments{
\item{cargs}{- Output from commandArgs(trailingOnly = F)}
}
\description{
When a script is being called from the command line using Rscript,
the relative path of the script being called will be present in the
command line arguments passed to the file. This is used to
get the relative path of the script.
}
