% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptloc.R
\name{scriptloc}
\alias{scriptloc}
\title{Get location of script in a fashion analagous to ${BASH_SOURCE[0]}}
\usage{
scriptloc()
}
\value{
Returns either a single string with path to the file being executed, or NULL
}
\description{
There are two ways in which code from an R file can be executed: through
the command line by invoking Rscript or by using the source() function.
This function tries to see if either of the methods were used, and if not,
deduces that the function is being called from an interactive session and
therefore returns NULL.
}
\examples{
writeLines("library(scriptloc); script_path <- scriptloc(); print(script_path)", "example.R")
source("example.R")
file.remove("example.R")
}
